+++ examples/serverless-otel/packages/functions/src/tracing.ts
@@ -4,9 +4,12 @@ import { AwsInstrumentation } from '@opentelemetry/instrumentation-aws-sdk';
 new BaselimeSDK({
     baselimeKey: process.env.BASELIME_KEY, collectorUrl: 'https://otel.baselime.cc/v1', serverless: false, instrumentations: [
       new AwsInstrumentation({
-        suppressInternalInstrumentation: false,
+        suppressInternalInstrumentation: true,
       }),
-      new BetterHttpInstrumentation()
+      new BetterHttpInstrumentation({ 
+        captureBody: true,
+        captureHeaders: true,
+      })
     ]
   }).start();